Update from use case review
It was determined through use that P code and disabling the D port would be useful additions.  There was also a s/w bug where the serial port uses D0/D1 for Rx/Tx and was causing issues with certain codes using those pins as LSBs.  Serial communication was removed from the device (it may be possible to add back if necessary) and port pins are set to output when port is enabled.